+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.2.0] - 2025-04-01
+### Added
+- Added feature of OpenAPI 3.0 schema auto-generation.
+- Added unit tests for the OpenAPI 3.0 schema auto-generation feature.
+
 ## [1.1.0] - 2025-03-17
 ### Added
 - Implemented dynamic routing (support for `GET users/:group/:limit` format)
@@ -19,5 +24,6 @@
 ---
 
 #### 🔗 GitHub Releases
+[1.2.0]: https://github.com/ka215/MockAPI-PHP/releases/tag/v1.2.0
 [1.1.0]: https://github.com/ka215/MockAPI-PHP/releases/tag/v1.1.0  
 [1.0.0]: https://github.com/ka215/MockAPI-PHP/releases/tag/v1.0.0